Update app.js
This is a script which contains all of the code for the basic bot. It also contains comments that explain what specific lines of code do. Replace the bot login token text with the token that was generated when creating your bot (refer to https://discordapp.com/developers/ for information regarding creating bots).